requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
dj-database-url==0.5.0
Django==2.0.3
django-crispy-forms==1.7.2
djangorestframework==3.10.2
gunicorn==19.9.0
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
mysqlclient==1.4.2.post1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.2
requests==2.22.0
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
whitenoise==4.1.3
